Refactor holiday models
