--- Models/HolidayApiResponse.cs
@@ -1,10 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace Api.Models
-{
-  public class HolidayApiResponse
-  {
-    [JsonPropertyName("holidays")]
-    public required Holiday[] Holidays { get; set; }
-  }
-}
